requirements.txt
python-pptx>=0.6.21
pillow>=9.0.0
cairosvg>=2.7.0
